Initial Commit
The main thing that I did here was that I removed tons of deprecated code and I also removed all of Artem's designs so we don't risk a lawsuit. The big idea behind NewzBytes is that we're going to impose a two minute or so limit on video legnth so we can keep things concise and not waste money on storage.